Rename corr2.py -> corr2base.py to disambiguate from treecorr.corr2 function
