+++ treecorr/corr2ex.py
@@ -16,7 +16,7 @@
 """
 
 from .catalog import Catalog, read_catalogs
-from .corr2 import Corr2
+from .corr2base import Corr2
 from .config import setup_logger, check_config, print_params, get
 from .util import set_omp_threads
 from .nncorrelation import NNCorrelation
