+++ treecorr/kkkcorrelation.py
@@ -19,7 +19,7 @@ import numpy as np
 
 from . import _treecorr as _lib
 from .catalog import calculateVarK
-from .corr3 import Corr3
+from .corr3base import Corr3
 from .util import make_writer, make_reader
 from .util import depr_pos_kwargs
