+++ treecorr/nncorrelation.py
@@ -18,7 +18,7 @@
 import numpy as np
 
 from . import _treecorr as _lib
-from .corr2 import Corr2
+from .corr2base import Corr2
 from .util import make_writer, make_reader, lazy_property
 from .util import depr_pos_kwargs
